requirements.txt
pillow==11.2.1
PyPDF2
anthropic
azure
beautifulsoup4
difflib
python-dotenv==1.1.0
google
langchain
langchain_community
langchain_core
langchain_experimental
langchain_google_genai
langchain_mistralai
pandas==2.2.3
pdfplumber
pydantic==2.11.4
pytesseract
requests==2.32.3
streamlit==1.44.1